second commit

+++ src/main/java/com/example/hra/service/LocationServiceImplement.java
@@ -5,6 +5,7 @@ import com.example.hra.Repository.LocationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.math.BigDecimal;
 import java.util.List;
 
@@ -19,15 +20,24 @@ public class LocationServiceImplement implements LocationService {
     }
 
     @Override
-    public String addLocation(Location location) {
+    public void addLocation(Location location) {
         locationRepository.save(location);
-        return "Record Created Successfully";
     }
 
     @Override
-    public String updateLocation(Location location) {
-        locationRepository.save(location);
-        return "Record Modified Successfully";
+    public void updateLocation(Location location) {
+        Location location1 = (Location) locationRepository.findByLocationId(location.getLocationId()).get();
+        if(location1!=null)
+        {
+            location1.setLocationId(location.getLocationId());
+            location1.setStreetAddress(location.getStreetAddress());
+            location1.setCity(location.getCity());
+            location1.setPostalCode(location.getPostalCode());
+            location1.setStateProvince(location.getStateProvince());
+            location1.setCountry(location.getCountry());
+            locationRepository.save(location);
+        }
+
     }
 
     @Override
@@ -42,9 +52,19 @@ public class LocationServiceImplement implements LocationService {
     }
 
     @Override
+    @Transactional
     public String deleteByLocationId(BigDecimal locationId) {
-        locationRepository.deleteByLocationId(locationId);
-        return "Record deleted Successfully";
+        Location location = (Location) locationRepository.findByLocationId(locationId).get();
+        if(location!=null)
+        {
+            locationRepository.deleteByLocationId(locationId);
+            return "Record deleted Successfully";
+        }
+        else
+        {
+            return "Record Not Found";
+        }
+
     }
 }
